+++ CHANGES.rst
@@ -203,6 +203,10 @@ Bug Fixes
 
 - ``astropy.io.ascii``
 
+  - When reading a table with values that generate an overflow error
+    during type conversion (e.g. overflowing the native C long type), fall
+    through to using string. Previously this generated an exception [#2234].
+
 - ``astropy.io.fits``
 
 - ``astropy.io.misc``